change: split niceQuest utils into their own file
